Fix bug in cProcess.foGetModule
If the debugger would die while enumerating modules, the
cProcess.fdoGetModules_by_sCdbId method returns None.
cProcess.foGetModule used to assume it would always return a dictionary.
The fix is to check if cdb is still running and return None if it is not
before attempting to retreive the module from the dictionary.