Do not ignore stack frames outside modules
Originally all stack frames that were not inside a known module where
assumed to be generated code (e.g. compiled Javascript). However, it
turns out there are situations where the stack is corrupt (or at least
WinDBG is unable to unwind it) which are nearly indistinguishable. One
might consider determining if the address in question points to
executable memory, but a corrupt stack could point to executable memory
by pure chance, making the stack hash unreliable.

In order to make the stack hash more reliable, stack hashing is
terminated when a stack frame outside of a known module is found. This
does mean the hash will be shorter and therefore stack collisions are
slightly more likely to occur. However, at this point I believe chances
of that happening are remote and increasing the stack hash length for
each frame could be used to counteract this.