Fix stack hash for AVE
* When hashing the stack, ignore hidden stack frames.

Previous edits stopped hashing stack when a stack frame with no id was
encountered (i.e. it is outside of a module). When an AVE happens, the
top stack frame will almost certainly be outside a module and have no
id. This lead to all AVE's having no stack hash. It makes more sense to
ignore all stack frames that have been hidden and only stop hashing when
a non-hidden stack frame with no id is encountered.