Improve stack exhaustion handling
* new dxBugIdConfig settings allow you to provide the minimum number of
loops and maximum number of functions in each loop that are used to
detect stack recursion in case of a stack exhaustion exception.
* Number of stack frames for stack exhaustion exceptions is now
dynamically increased to allow accommodate the minimum number of loops
and maximum number of functions in each loop (it used to be that it
might be too small to detect large loops).