Improve application runtime calculation
* Problem: code used to calculate application runtime based on the time
elapsed between starting/resuming the application and when cdb reports
an event. However, cdb can take quite a while to process an event, which
incorrectly got added to the application runtime.
* cdb returns a timestamp for each event, which (AFAIK) represents the
time the event was detected by cdb, rather than the time when it was
reported. This can be used to better calculate the time the application
has been running.
* The excessive CPU usage detector now uses the calculated application
runtime described above to determine CPU usage percentage, rather than
the thread run-time.

This patch also improves the following:
* New/terminated process ids are now output in decimal and hexadecimal.
* More code executed while a lock is held is now put inside a try ...
finally block to make sure the lock is released in case of an exception.
* Debug output for excessive CPU usage should now be easier to
understand.