Attempt to resolve an issue on Windows 7
With GFlags set to a string containing a quote ("), an error can occur
on Windows 7 because the entire IF block would get parsed and %GFlags%
expanded to the quoted string.

The script was designed to have %GFlags% be expanded later, after the
variable has been set. To accomplish this, the contents of the IF block
have been moved to a separate section of the batch script, which is
called by its label. This should cause the script lines containing
%GFlags% to be interpreted only if GFlags was not set by the user and
after GFlags has been set by the script to a string that does not
contain any quotes.