Delayed path creation for Firefox profile
Static arguments are created as needed now through a function call, and
no longer from the main code. This allows the code to create the Firefox
profile folder only when it is needed and always return a 8.3 path to
that folder. Previously, it was impossible to provide a 8.3 path if the
folder did not exist, and you would not want to always create the
Firefox profile path if you were not debugging Firefox...